recomendation servic/requirements.txt
# FastAPI
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0

# Database
SQLAlchemy==1.4.52
psycopg2-binary==2.9.7
alembic==1.12.0

# Pydantic
pydantic==1.10.7

# ML Libraries
catboost==1.2
gensim==4.3.2
scikit-learn==1.3.0

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Serialization
joblib==1.3.2